Added the registration reset button and it's functionality.

+++ public/pages/register.php
@@ -47,10 +47,8 @@ function set_title($title = null)
     <main class="container py-4">
         <div class="form-container text-center">
             <div class="form-top-container">
-                <svg class="header-back-button cursor-pointer" width="18" height="15" viewBox="0 0 16 15" fill="none" xmlns="http://www.w3.org/2000/svg">
-                    <path fill-rule="evenodd" clip-rule="evenodd"
-                        d="M8.60564 1.65147L3.73182 6.5253H16V8.47483H3.73182L8.60564 13.3487L7.22712 14.7272L0 7.50006L7.22712 0.272949L8.60564 1.65147Z"
-                        fill="#722F37" />
+                <svg id="reset-registration" class="cursor-pointer" width="22" height="22" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+                    <path d="M21.1679 8C19.6248 4.46819 16.1006 2 12 2C6.81465 2 2.5511 5.94668 2.04938 11M22 3V7.4C22 7.73137 21.7314 8 21.4 8H17M2.88146 16C4.42458 19.5318 7.94874 22 12.0494 22C17.2347 22 21.4983 18.0533 22 13M2.04932 21V16.6C2.04932 16.2686 2.31795 16 2.64932 16H7.04932" stroke="#94241E" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round" />
                 </svg>
                 <div class="pagination">
                     <span class="page active"></span>
@@ -86,9 +84,14 @@ function set_title($title = null)
                     </div>
                     <div class="form-field">
                         <input type="email" name="email" id="email" placeholder="Email address" class="input">
-                        <button type="button" class="btn primary-btn mt-3" id="email-submit">
-                            Continue
-                        </button>
+                        <div class="d-flex justify-content-between">
+                            <button type="button" class="btn bg-light mt-3 prev-button">
+                                Back
+                            </button>
+                            <button type="button" class="btn primary-btn mt-3" id="email-submit">
+                                Continue
+                            </button>
+                        </div>
                     </div>
                 </div>
                 <!-- OTP VERIFICATION -->
@@ -130,9 +133,14 @@ function set_title($title = null)
                             <input type="text" id="phone" name="phone_number" maxlength="10" placeholder="Phone Number"
                                 class="input">
                         </div>
-                        <button type="button" class="btn primary-btn mt-3" id="phone-submit">
-                            Continue
-                        </button>
+                        <div class="d-flex justify-content-between">
+                            <button type="button" class="btn bg-light mt-3 prev-button">
+                                Back
+                            </button>
+                            <button type="button" class="btn primary-btn mt-3" id="phone-submit">
+                                Continue
+                            </button>
+                        </div>
                     </div>
                 </div>
 
@@ -180,9 +188,29 @@ function set_title($title = null)
 </body>
 <script>
     // Reset registration
-    document.getElementById('registration_reset').addEventListener('click', function() {
-        sessionStorage.clear();
-        window.location.href = 'register.php';
+    document.addEventListener('DOMContentLoaded', function() {
+        const resetBtn = document.getElementById('reset-registration');
+        resetBtn.addEventListener('click', function() {
+            if (sessionStorage.getItem('registration_id')) {
+                const registration_id = sessionStorage.getItem('registration_id');
+                sendAjaxRequest(
+                    "reset-registration.php",
+                    "POST",
+                    "registration_id=" + encodeURIComponent(registration_id),
+                    function(response) {
+                        if (!response.success) {
+                            showToasted(response.message, "error");
+                        } else {
+                            showToasted(response.message, "success");
+                            setTimeout(() => {
+                                sessionStorage.clear();
+                                window.location.href = "register.php";
+                            }, 2000);
+                        }
+                    }
+                );
+            }
+        });
     });
 
     // Change referral submit button text based on input
